set up the Blueprint in the auth.py - g.user
