+++ portal/static/styles.css
@@ -60,5 +60,3 @@ label {
 label input {
   width: 100%;
 }
-
-
